src/requirements/core.txt
beautifulsoup4==4.4.1
bleach==1.4.3
Cartridge==0.11.0
chardet==2.3.0
Django==1.9.7
django-contrib-comments==1.7.1
django-haystack==2.4.1
django-machina==0.2.1
django-markdown==0.8.4
django-modeltranslation==0.11
django-mptt==0.8.4
django-widget-tweaks==1.4.1
filebrowser-safe==0.4.3
future==0.15.2
grappelli-safe==0.4.2
html5lib==0.9999999
Markdown==2.6.6
Mezzanine==4.1.0
oauthlib==1.1.2
Pillow==3.2.0
psycopg2==2.6.1
PyPDF2==1.26.0
pytz==2016.4
reportlab==3.3.0
requests==2.10.0
requests-oauthlib==0.6.1
six==1.10.0
tzlocal==1.2.2
xhtml2pdf==0.0.6
